requirements.txt
numpy
pandas
matplotlib
seaborn
pyyaml
jupyter
ipykernel
